Add files via upload
This is our final assignment for computer graphic subject. Our theme is a camp site at night. In this project, we put several object in the scene so it looks more realistic and immersive. The object consist of a person, a campfire, a tent and a pine tree. To make it look even more immersive, we put our scene inside a cube and we apply a skybox using cubemapping. We apply a same style as a free camera, camera so we can look around our scene